adapting code for passing precommit
