--- exec_manager/JobFactory.py
@@ -1,9 +1,0 @@
-import ExecProfile
-import Job
-
-class JobFactory:
-    def __init__(self) -> None:
-        pass
-
-    def create_job(self, inputs: dict, workflow, exec_profile: ExecProfile) -> Job:
-        pass
