--- exec_manager/dao/JobDAO.py
@@ -1,9 +1,0 @@
-from uuid import uuid4
-import JobStatusType
-
-class JobDAO:
-    def __init__(self) -> None:
-        pass
-
-    def update_job_status(self, job_id: uuid4, new_job_status: JobStatusType) -> None:
-        pass
